Automatically deploy entrypoint, and make it available in the hre
